+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh61ggd195yf4t61s8f9hwnq"
+        "conv_01kh6d09ffqr9qaezvpdcbdxst"
     )
     r"""Unique conversation identifier with `conv_` prefix."""