+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-11T09:46:14.324Z")
+    created: Optional[datetime] = parse_datetime("2026-02-11T13:07:06.124Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")