+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH61GGX566B62WNWS112VKG9"
+        "01KH6D0A2JDBZJDMP5V1XFYH6Y"
     )
     r"""The unique identifier of the data source"""