+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH61GGX7JSQ66RCT7Z5EQHQA"
+        "01KH6D0A2MVWJPP7TRG1M39VN5"
     )
     r"""The unique identifier of the data source"""