+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH61GGXCE0J743S8R61V9AVN"
+        "01KH6D0A30VRMF4F41K3E66N9H"
     )
     r"""The unique identifier of the data source"""