+++ packages/orq-rc/RELEASES.md
@@ -5519,3 +5519,13 @@ Based on:
 - [python v4.3.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.52 - packages/orq-rc
+
+## 2026-02-11 13:09:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.710.0 (2.818.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.53 - packages/orq-rc